isolate DataMode/DataUnit, FileMetadata, file system processing funcs
