+++ summawise/serializable.py
@@ -1,9 +1,9 @@
 import json
-from .utils import DataMode
-from .files import utils as FileUtils
 from typing import TypeVar, Type
 from pathlib import Path
 from dataclasses import is_dataclass, asdict
+from .data import DataMode
+from .files import utils as FileUtils
 
 ST = TypeVar("ST", bound = "Serializable")
